requirements.txt
Flask
Flask-Login
Werkzeug
requests
google-genai
gunicorn
flask-sqlalchemy
